[platform_device_loader] isolate archive related operations
