+++ libparsec/crates/platform_device_loader/src/device/mod.rs
@@ -1,0 +1,5 @@
+// Parsec Cloud (https://parsec.cloud) Copyright (c) BUSL-1.1 2016-present Scille SAS
+
+mod archive;
+
+pub use archive::*;
